move MemoryMessageStream to own file
